requirements.txt
<<<<<<< HEAD
django==4.2.13
djangorestframework==3.14.0 
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
python-decouple==3.8
django-cors-headers==4.3.1
social-auth-app-django==5.4.0
requests==2.31.0
pytest==7.4.3
pytest-django==4.5.2
=======
asgiref==3.9.1
Django==5.2.4
sqlparse==0.5.3
tzdata==2025.2
>>>>>>> ft-history
